+++ src/options.ts
@@ -138,7 +138,7 @@ export class Options {
       let contents: string[] = [];
       let currentSection = '';
 
-      const found = {};
+      const found: Record<string, boolean> = {};
       let lines = content.split('\n');
       for (var i = 0; i < lines.length; i++) {
         let line = lines[i];
@@ -309,6 +309,23 @@ export class Options {
     return vscode.workspace.getConfiguration().get('wakatime.apiUrl') || '';
   }
 
+  public getStatusBarAlignment(): vscode.StatusBarAlignment {
+    const align: string = vscode.workspace.getConfiguration().get('wakatime.align') ?? '';
+    switch (align) {
+      case 'left':
+        return vscode.StatusBarAlignment.Left;
+      case 'right':
+        return vscode.StatusBarAlignment.Right;
+      default:
+        return vscode.StatusBarAlignment.Left;
+    }
+  }
+
+  public getStatusBarPriority(): number {
+    const priority = vscode.workspace.getConfiguration().get('wakatime.alignPriority');
+    return typeof priority === 'number' ? priority : 1;
+  }
+
   // Support for gitpod.io https://github.com/wakatime/vscode-wakatime/pull/220
   public getApiKeyFromEnv(): string {
     if (this.cache.api_key_from_env !== undefined) return this.cache.api_key_from_env;
@@ -328,6 +345,9 @@ export class Options {
     if (!apiUrl && !checkSettingsFile) {
       return '';
     }
+
+    // people often accidentally enter their API Key into the API Url settings input
+    if (!Utils.validateApiUrl(apiUrl)) apiUrl = '';
 
     if (!apiUrl) {
       try {